03/pyproject.toml
[project]
name = "03"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "mysql-connector-python>=9.5.0",
    "python-dotenv>=1.2.1",
]
